Move top page styles to app dir
