+++ app/style.module.scss
@@ -1,4 +1,4 @@
-@use 'grid';
+@use '../styles/top-page/grid';
 
 .main-window {
     margin: 1em;
