+++ app/toppage/Ratings.module.scss
@@ -1,4 +1,4 @@
-@use '../mixins';
+@use '../../styles/mixins';
 
 .rating_list {
     font-weight: bold;
