+++ app/toppage/TopPageBalloons.tsx
@@ -1,5 +1,5 @@
 import Block from "../../components/Block";
-import styles from "../../styles/top-page/main.module.scss";
+import styles from "../style.module.scss";
 import Link from "next/link";
 import dynamic from "next/dynamic";
