+++ app/toppage/WhatsNew.module.scss
@@ -1,4 +1,4 @@
-@use '../mixins';
+@use '../../styles/mixins';
 
 #whats_new_table {
     overflow-y: scroll;
