+++ app/toppage/WhatsNew.tsx
@@ -1,4 +1,4 @@
-import styles from "../../styles/top-page/WhatsNew.module.scss";
+import styles from "./WhatsNew.module.scss";
 import ReactMarkdown from "react-markdown";
 import Block from "../../components/Block";
 import {getWhatsNewRecords, WhatsNewRecord} from "../../lib/whats_new";
